+++ src/js/__tests__/sdk.ts
@@ -2,6 +2,7 @@ import { getGlobalObject } from '@sentry/utils';
 
 import type { CordovaOptions } from '../options';
 import * as Sdk from '../sdk';
+import { getPlatform } from '../utils';
 
 const optionsTest: {
   current?: CordovaOptions;
@@ -16,6 +17,15 @@ jest.mock('@sentry/core', () => {
     ...core,
     initAndBind: jest.fn((_, options) => (optionsTest.current = options)),
     makeMain: jest.fn(),
+  };
+});
+
+jest.mock('../utils', () => {
+  const util = jest.requireActual('../utils');
+
+  return {
+    ...util,
+    getPlatform: jest.fn().mockReturnValue('android'),
   };
 });
 
@@ -44,5 +54,42 @@ describe('Tests SDK', () => {
 
       expect(optionsTest.current?.release).toBe('user-release');
     });
+
+    describe('ios Options', () => {
+
+      it('Should include iOS parameters when running on iOS', async () => {
+        (getPlatform as jest.Mock).mockReturnValue('ios');
+
+        const expectedOptions: CordovaOptions = {
+          environment: 'abc',
+          // iOS parameters
+          enableAppHangTracking: true,
+          appHangTimeoutInterval: 123
+        };
+
+        Sdk.init(expectedOptions);
+
+        expect(optionsTest.current?.appHangTimeoutInterval).toEqual(expectedOptions.appHangTimeoutInterval);
+        expect(optionsTest.current?.enableAppHangTracking).toEqual(expectedOptions.enableAppHangTracking);
+      });
+
+      it('Should not include iOS parameters when running on android', async () => {
+        (getPlatform as jest.Mock).mockReturnValue('android');
+
+        const expectedOption = {
+          environment: 'abc'
+        }
+        const unexpectedOptions = {
+          appHangTimeoutInterval: 123,
+          enableAppHangTracking: true
+        };
+
+        Sdk.init({ ...unexpectedOptions, ...expectedOption });
+
+        expect(optionsTest.current).not.toContain(unexpectedOptions);
+        expect(optionsTest.current?.environment).toEqual(expectedOption.environment);
+      });
+    })
+
   });
 });